@@ -1,7 +1,7 @@
 [package]
 name = "mevlog"
 version = "0.7.0"
-edition = "2021"
+edition = "2024"
 repository = "https://github.com/pawurb/mevlog-rs"
 description = "EVM transactions monitoring and querying CLI powered by Revm"
 readme = "README.md"
